Let's not use long - at least not publicly
It was probably important when there was such thing as 16-bit system.

+++ include/grx/color.h
@@ -146,7 +146,7 @@ extern const struct _GR_colorInfo {
         unsigned char r,g,b;            /* loaded components */
         unsigned int  defined:1;        /* r,g,b values are valid if set */
         unsigned int  writable:1;       /* can be changed by 'grx_color_cell_set' */
-        unsigned long int nused;        /* usage count */
+        unsigned int  nused;            /* usage count */
     } ctable[256];
 } * const GrColorInfo;
